backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
celery[redis]
redis
httpx
google-genai
pgvector
pydantic-settings
psutil
psycopg[binary]
